backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
GeoAlchemy2==0.14.2

# Computer Vision & AI
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
scikit-image==0.22.0
numpy==1.24.4

# Geospatial Processing
Shapely==2.0.2
Fiona==1.9.5
rasterio==1.3.9
pyproj==3.6.1
folium==0.15.0

# API & Utilities
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==3.0.1
gunicorn==21.2.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
geopandas==0.14.1
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0